requirements.txt
autopep8==1.4.3
chainer==5.0.0
chainercv==0.11.0
cupy-cuda91==5.0.0
cycler==0.10.0
dill==0.2.8.2
fastrlock==0.4
filelock==3.0.10
flake8==3.6.0
greenlet==0.4.15
joblib==0.13.0
kiwisolver==1.0.1
matplotlib==3.0.2
mccabe==0.6.1
msgpack==0.5.6
multiprocess==0.70.6.1
neovim==0.3.1
numpy==1.15.4
opencv-python==3.4.4.19
pathlib==1.0.1
Pillow==5.3.0
protobuf==3.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pynvim==0.3.1
pyparsing==2.3.0
python-dateutil==2.7.5
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tqdm==4.28.1
